+++ catalog/controller/payment/nuvei.php
@@ -92,8 +92,8 @@ class Nuvei extends \Opencart\System\Engine\Controller
             'showResponseMessage'       => false,
             'sessionToken'              => $order_data['sessionToken'],
             'env'                       => 1 == $test_mode ? 'test' : 'prod',
-            'merchantId'                => $this->plugin_settings[NUVEI_SETTINGS_PREFIX . 'merchantId'],
-            'merchantSiteId'            => $this->plugin_settings[NUVEI_SETTINGS_PREFIX . 'merchantSiteId'],
+            'merchantId'                => trim($this->plugin_settings[NUVEI_SETTINGS_PREFIX . 'merchantId']),
+            'merchantSiteId'            => trim($this->plugin_settings[NUVEI_SETTINGS_PREFIX . 'merchantSiteId']),
             'country'                   => $order_data['billingAddress']['country'],
             'currency'                  => $order_data['currency'],
             'amount'                    => $order_data['amount'],
@@ -750,19 +750,6 @@ class Nuvei extends \Opencart\System\Engine\Controller
                                                     = NUVEI_SDK_AUTOCLOSE_URL;
         }
         
-        # use or not UPOs
-        // in case there is a Product with a Payment Plan
-//        if (!empty($rebilling_params['merchantDetails']['customField3'])) {
-//            $oo_params['userTokenId'] = $oo_params['billingAddress']['email'];
-//        }
-//        if(isset($this->plugin_settings[NUVEI_SETTINGS_PREFIX . 'use_upos'])
-//            && 1 == $this->plugin_settings[NUVEI_SETTINGS_PREFIX . 'use_upos'] 
-//            && 1 == $this->is_user_logged
-//        ) {
-//            $oo_params['userTokenId'] = $oo_params['billingAddress']['email'];
-//        }
-        # /use or not UPOs
-        
         $oo_params = array_merge_recursive($oo_params, $rebilling_params);
         
 		$resp = \Nuvei_Class::call_rest_api(
@@ -829,7 +816,7 @@ class Nuvei extends \Opencart\System\Engine\Controller
 		if (!empty($advanceResponseChecksum)) {
             $str = hash(
                 $this->config->get(NUVEI_SETTINGS_PREFIX . 'hash'),
-                $this->config->get(NUVEI_SETTINGS_PREFIX . 'secret')
+                trim($this->config->get(NUVEI_SETTINGS_PREFIX . 'secret'))
                     . \Nuvei_Class::get_param('totalAmount')
                     . \Nuvei_Class::get_param('currency')
                     . \Nuvei_Class::get_param('responseTimeStamp')
@@ -870,7 +857,7 @@ class Nuvei extends \Opencart\System\Engine\Controller
 		$dmn_params = array_diff_key($request_arr, $custom_params);
 		$concat     = implode('', $dmn_params);
 		
-		$concat_final = $concat . $this->config->get(NUVEI_SETTINGS_PREFIX . 'secret');
+		$concat_final = $concat . trim($this->config->get(NUVEI_SETTINGS_PREFIX . 'secret'));
 		$checksum     = hash($this->config->get(NUVEI_SETTINGS_PREFIX . 'hash'), $concat_final);
 		
 		if ($responsechecksum !== $checksum) {