+++ nuvei_class.php
@@ -189,8 +189,8 @@ class Nuvei_Class
         // set here some of the mandatory parameters
         $params = array_merge_recursive(
             array(
-                'merchantId'        => $settings[NUVEI_SETTINGS_PREFIX . 'merchantId'],
-                'merchantSiteId'    => $settings[NUVEI_SETTINGS_PREFIX . 'merchantSiteId'],
+                'merchantId'        => trim($settings[NUVEI_SETTINGS_PREFIX . 'merchantId']),
+                'merchantSiteId'    => trim($settings[NUVEI_SETTINGS_PREFIX . 'merchantSiteId']),
                 'timeStamp'         => $time,
                 'webMasterId'       => 'OpenCart ' . VERSION . '; Plugin v' . self::get_plugin_version(),
                 'sourceApplication' => NUVEI_SOURCE_APP,
@@ -225,7 +225,7 @@ class Nuvei_Class
             $concat .= $params[$key];
         }
         
-        $concat .= $settings[NUVEI_SETTINGS_PREFIX . 'secret'];
+        $concat .= trim($settings[NUVEI_SETTINGS_PREFIX . 'secret']);
         
         $params['checksum'] = hash($settings[NUVEI_SETTINGS_PREFIX . 'hash'], $concat);
         // /calculate the checksum